back-end/requirements.txt
asgiref==3.11.0
Django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.3.0
sqlparse==0.5.5
tzdata==2025.3
PyPDF2==3.0.1
python-docx==0.8.11

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.2
django-filter==24.1

# Task Queue & Caching
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Data Processing & ML
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2

# Utilities
python-decouple==3.8
Pillow==10.1.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
